% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_multstart.R
\name{nls_multstart}
\alias{nls_multstart}
\title{Finds the best fit of non-linear model based on AIC score}
\usage{
nls_multstart(formula, data = parent.frame(), iter, param_bds,
  supp_errors = c("Y", "N"), AICc = c("Y", "N"), control, ...)
}
\arguments{
\item{formula}{a non-linear model formula, with the response on the left of a ~ operator and an expression involving parameters on the right}

\item{data}{data.frame (optional) in which to evaluate the variables in \code{formula} and \code{weights}}

\item{iter}{number of combinations of starting parameters that
are tried on each curve.}

\item{param_bds}{lower and upper boundaries for the start parameters. If
missing these default to +/- 1e+09. Need to specified as a vector as :
c(lower bound param 1, upper bound param 1, lower bound param 2, upper bound
param 2 etc)}

\item{supp_errors}{if \code{supp_errors = 'Y'}, then no error messages will be shown
from the nlsLM function, reducing the number of error messages
printed while the model attempts to converge using poor starting parameters. Advised to only use \code{supp_errors = 'Y'} when you are confident in the bounds of your starting parameters.}

\item{AICc}{whether or not the small sample AIC should be used. Defaults to
\code{'Y'}. Override this using \code{AICc == 'N'}. AICc should be used instead of AIC
when sample size is small in comparison to the number of estimated
parameters (Burnham & Anderson 2002 recommend its use when n / n_param < 40).}

\item{control}{specific control can be specified using \code{\link[minpack.lm]{nls.lm.control}}.}

\item{\dots}{Extra arguments to pass to \code{\link[minpack.lm]{nlsLM}} if necessary.}
}
\value{
returns a nls object of the best estimated model fit
}
\description{
Finds the best estimated model using non-linear least squares regression using nlsLM(). The best fit is determined using AIC scores.
}
\note{
Useful additional arguments for \code{\link[minpack.lm]{nlsLM}} include: \code{na.action = na.omit},
\code{lower/upper = c()} where these represent upper and lower boundaries for parameter estimates
}
\examples{
# load in data

data("Chlorella_TRC")
Chlorella_TRC_test <- Chlorella_TRC[Chlorella_TRC$curve_id == 1,]

# run nlsLoop()

# define the Sharpe-Schoolfield equation
schoolfield_high <- function(lnc, E, Eh, Th, temp, Tc) {
 Tc <- 273.15 + Tc
 k <- 8.62e-5
 boltzmann.term <- lnc + log(exp(E/k*(1/Tc - 1/temp)))
 inactivation.term <- log(1/(1 + exp(Eh/k*(1/Th - 1/temp))))
 return(boltzmann.term + inactivation.term)
}

fits <- nls_multstart(ln.rate ~ schoolfield_high(lnc, E, Eh, Th, temp = K, Tc = 20),
                data = Chlorella_TRC_test,
                iter = 500,
                param_bds = c(-10, 10, 0.1, 2, 0.5, 5, 285, 330),
                lower = c(lnc=-10, E=0, Eh=0, Th=0),
                supp_errors = 'Y')

}
\seealso{
\code{\link[minpack.lm]{nlsLM}} for details on additional arguments to pass to the nlsLM function.

\code{\link[MuMIn]{AICc}} for application of AICc.
}
\author{
Daniel Padfield
}
